chore(ci): make CircleCI perform the deploy step upon tagged commit on master
This commit adds an extra step to the CircleCI configuration to release
the project to PyPI if and only if the CI is passing on a semver
tagged commit from master and all tests are passing.

In essence, it prevents the releaser to locally configure PyPI and
fetch credentials and let CircleCI perform the `twine upload` on its own.